Convert existing pages to templates
