+++ @types/types.d.ts
@@ -1,0 +1,3 @@
+export type SupportedLocales = 'en' | 'ja'
+
+export type TopLevelTemplate = (props: {locale: SupportedLocales}) => JSX.Element;
